refactor: extract search sql engine
