+++ src/core/src/service/search/utils.rs
@@ -19,10 +19,10 @@ use std::{
     sync::{Arc, atomic::Ordering},
 };
 
+pub use ::search::utils::{conjunction, is_field, is_value, split_conjunction, trim_quotes};
 use config::meta::search::{PARTIAL_ERROR_RESPONSE_MESSAGE, ScanStats};
 use datafusion::physical_plan::{ExecutionPlan, ExecutionPlanVisitor};
 use infra::runtime::DATAFUSION_RUNTIME;
-use sqlparser::ast::{BinaryOperator, Expr};
 use tokio::sync::Mutex;
 
 use super::datafusion::distributed_plan::remote_scan_exec::RemoteScanExec;
@@ -103,87 +103,6 @@ impl ExecutionPlanVisitor for ScanStatsVisitor {
         }
         Ok(true)
     }
-}
-
-// split an expression by AND operator
-pub fn split_conjunction(expr: &Expr) -> Vec<&Expr> {
-    split_conjunction_inner(expr, Vec::new())
-}
-
-fn split_conjunction_inner<'a>(expr: &'a Expr, mut exprs: Vec<&'a Expr>) -> Vec<&'a Expr> {
-    match expr {
-        Expr::BinaryOp {
-            left,
-            op: BinaryOperator::And,
-            right,
-        } => {
-            let exprs = split_conjunction_inner(left, exprs);
-            split_conjunction_inner(right, exprs)
-        }
-        Expr::Nested(expr) => split_conjunction_inner(expr, exprs),
-        other => {
-            exprs.push(other);
-            exprs
-        }
-    }
-}
-
-// conjunction the exprs
-pub fn conjunction(exprs: Vec<&Expr>) -> Option<Expr> {
-    if exprs.is_empty() {
-        None
-    } else if exprs.len() == 1 {
-        Some(exprs[0].clone())
-    } else {
-        // conjuction all expr in exprs
-        let mut expr = exprs[0].clone();
-        if matches!(
-            expr,
-            Expr::BinaryOp {
-                op: BinaryOperator::Or,
-                ..
-            }
-        ) {
-            expr = Expr::Nested(Box::new(expr));
-        }
-        for e in exprs.into_iter().skip(1) {
-            expr = Expr::BinaryOp {
-                left: Box::new(expr),
-                op: BinaryOperator::And,
-                right: if matches!(
-                    e,
-                    Expr::BinaryOp {
-                        op: BinaryOperator::Or,
-                        ..
-                    }
-                ) {
-                    Box::new(Expr::Nested(Box::new(e.clone())))
-                } else {
-                    Box::new(e.clone())
-                },
-            }
-        }
-        Some(expr)
-    }
-}
-
-pub fn trim_quotes(s: &str) -> String {
-    let s = s
-        .strip_prefix('"')
-        .and_then(|s| s.strip_suffix('"'))
-        .unwrap_or(s);
-    s.strip_prefix('\'')
-        .and_then(|s| s.strip_suffix('\''))
-        .unwrap_or(s)
-        .to_string()
-}
-
-pub fn is_value(e: &Expr) -> bool {
-    matches!(e, Expr::Value(_))
-}
-
-pub fn is_field(e: &Expr) -> bool {
-    matches!(e, Expr::Identifier(_) | Expr::CompoundIdentifier(_))
 }
 
 pub fn check_query_default_limit_exceeded(num_rows: usize, partial_err: &mut String, sql: &Sql) {
