+++ src/search/Cargo.toml
@@ -7,18 +7,27 @@ publish = false
 
 [features]
 default = []
-enterprise = []
+enterprise = ["dep:o2_enterprise"]
 
 [dependencies]
+anyhow.workspace = true
+arrow-schema.workspace = true
 async-trait.workspace = true
+chrono.workspace = true
 config.workspace = true
+datafusion.workspace = true
+hashbrown.workspace = true
+infra.workspace = true
+log.workspace = true
+o2_enterprise = { workspace = true, optional = true, default-features = false }
+proto.workspace = true
+regex.workspace = true
 serde.workspace = true
+sqlparser.workspace = true
 utoipa.workspace = true
 vector-enrichment.workspace = true
 vrl.workspace = true
 
 [dev-dependencies]
 arrow.workspace = true
-arrow-schema.workspace = true
-o2_enterprise.workspace = true
 serde_json.workspace = true
