+++ src/search/src/lib.rs
@@ -14,6 +14,9 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 pub mod enrichment;
+pub mod sql;
+pub mod udf;
+pub mod utils;
 
 use config::meta::{
     search::{Response, ScanStats},
