+++ src/search/src/sql/rewriter/add_ordering_term.rs
