+++ src/search/src/sql/rewriter/track_total_hits.rs
