+++ src/search/src/sql/visitor/histogram_interval.rs
