+++ src/search/src/sql/visitor/pickup_where.rs
