+++ src/search/src/sql/visitor/utils.rs
