+++ src/search/src/udf.rs
@@ -1,0 +1,29 @@
+// Copyright 2026 OpenObserve Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+pub const DECRYPT_SLOW_UDF_NAME: &str = "decrypt";
+pub const DECRYPT_UDF_NAME: &str = "decrypt_path";
+pub const ENCRYPT_UDF_NAME: &str = "encrypt";
+pub const FUZZY_MATCH_ALL_UDF_NAME: &str = "fuzzy_match_all";
+pub const FUZZY_MATCH_UDF_NAME: &str = "fuzzy_match";
+pub const MATCH_ALL_HASH_UDF_NAME: &str = "match_all_hash";
+pub const MATCH_ALL_UDF_NAME: &str = "match_all";
+pub const MATCH_FIELD_IGNORE_CASE_UDF_NAME: &str = "match_field_ignore_case";
+pub const MATCH_FIELD_UDF_NAME: &str = "match_field";
+pub const REGEX_MATCHES_UDF_NAME: &str = "re_matches";
+pub const REGEX_MATCH_UDF_NAME: &str = "re_match";
+pub const REGEX_NOT_MATCH_UDF_NAME: &str = "re_not_match";
+pub const STR_MATCH_UDF_IGNORE_CASE_NAME: &str = "str_match_ignore_case";
+pub const STR_MATCH_UDF_NAME: &str = "str_match";
